fix: refact gh actions (#27)
* fix ci

fix: request changes

add cargo cache

add cargo cache

* cache cargo dir

replace runner machine

* lint code

* update executor rev

* adjust upgrader&releases

* update contract builds

* fix template tests

* fix rustup misspelling

* fix some scenarios tests

* invalid function code result && mock new address current nonce

* update Float Glob error

* remove forwarder tests in composability scenarios

* fix contract code missing upgrade entry

* erc20 proxy build config

* ignore cferc20 go tests due execute_on_dest_context not propagated in go

* max gas limit

* no log for TF with 0 value

* KDA error message for no payments && allocator in pair contract

* default value for nft transfer 1

* burn trigger mock

* NFT Mint && Update Metadata

* fix test error message

* contracts and stacking modules

* lint format

* clippy warnings

* Maximize build space

* fix contract register rs_tests

* Maximize disk space && template tests && coderabbitai suggestions

---------

+++ framework/base/src/types/interaction/contract_call_legacy.rs
@@ -18,7 +18,7 @@ pub use contract_call_with_multi_kda::ContractCallWithMultiKda;
 pub use contract_deploy::{new_contract_deploy, ContractDeploy};
 pub use typed_function_call::TypedFunctionCall;
 
-/// Using max u64 to represent maximum possible gas,
+/// Using max i64 to represent maximum possible gas,
 /// so that the value zero is not reserved and can be specified explicitly.
 /// Leaving the gas limit unspecified will replace it with `api.get_gas_left()`.
-pub(crate) const UNSPECIFIED_GAS_LIMIT: u64 = u64::MAX;
+pub(crate) const UNSPECIFIED_GAS_LIMIT: u64 = i64::MAX as u64;
